+++ src/components/block-tunes/block-tune-move-up.ts
@@ -117,6 +117,8 @@ export default class MoveUpTune implements BlockTune {
     /** Change blocks positions */
     this.api.blocks.move(currentBlockIndex - 1);
 
+    this.api.toolbar.toggleBlockSettings(true);
+
     /** Hide the Tooltip */
     this.api.tooltip.hide();
   }