+++ src/components/modules/blockEvents.ts
@@ -125,9 +125,10 @@ export default class BlockEvents extends Module {
       return;
     }
 
-    const canOpenToolbox = currentBlock.tool.isDefault && currentBlock.isEmpty;
-    const conversionToolbarOpened = !currentBlock.isEmpty && ConversionToolbar.opened;
-    const inlineToolbarOpened = !currentBlock.isEmpty && !SelectionUtils.isCollapsed && InlineToolbar.opened;
+    const isEmptyBlock = currentBlock.isEmpty;
+    const canOpenToolbox = currentBlock.tool.isDefault && isEmptyBlock;
+    const conversionToolbarOpened = !isEmptyBlock && ConversionToolbar.opened;
+    const inlineToolbarOpened = !isEmptyBlock && !SelectionUtils.isCollapsed && InlineToolbar.opened;
 
     /**
      * For empty Blocks we show Plus button via Toolbox only for default Blocks
@@ -255,19 +256,9 @@ export default class BlockEvents extends Module {
     this.Editor.Caret.setToBlock(newCurrent);
 
     /**
-     * If new Block is empty
+     * Show Toolbar
      */
-    if (newCurrent.tool.isDefault && newCurrent.isEmpty) {
-      /**
-       * Show Toolbar
-       */
-      this.Editor.Toolbar.open(false);
-
-      /**
-       * Show Plus Button
-       */
-      this.Editor.Toolbar.plusButton.show();
-    }
+    this.Editor.Toolbar.moveAndOpen(newCurrent);
 
     event.preventDefault();
   }
@@ -505,7 +496,7 @@ export default class BlockEvents extends Module {
    * @param {KeyboardEvent} event - keyboard event
    */
   private needToolbarClosing(event: KeyboardEvent): boolean {
-    const toolboxItemSelected = (event.keyCode === _.keyCodes.ENTER && this.Editor.Toolbox.opened),
+    const toolboxItemSelected = (event.keyCode === _.keyCodes.ENTER && this.Editor.Toolbar.toolbox.opened),
         blockSettingsItemSelected = (event.keyCode === _.keyCodes.ENTER && this.Editor.BlockSettings.opened),
         inlineToolbarItemSelected = (event.keyCode === _.keyCodes.ENTER && this.Editor.InlineToolbar.opened),
         conversionToolbarItemSelected = (event.keyCode === _.keyCodes.ENTER && this.Editor.ConversionToolbar.opened),
@@ -531,11 +522,10 @@ export default class BlockEvents extends Module {
    */
   private activateToolbox(): void {
     if (!this.Editor.Toolbar.opened) {
-      this.Editor.Toolbar.open(false, false);
-      this.Editor.Toolbar.plusButton.show();
-    }
+      this.Editor.Toolbar.moveAndOpen();
+    } // else Flipper will leaf through it
 
-    this.Editor.Toolbox.open();
+    this.Editor.Toolbar.toolbox.open();
   }
 
   /**
@@ -544,8 +534,7 @@ export default class BlockEvents extends Module {
   private activateBlockSettings(): void {
     if (!this.Editor.Toolbar.opened) {
       this.Editor.BlockManager.currentBlock.focused = true;
-      this.Editor.Toolbar.open(true, false);
-      this.Editor.Toolbar.plusButton.hide();
+      this.Editor.Toolbar.moveAndOpen();
     }
 
     /**
@@ -553,6 +542,11 @@ export default class BlockEvents extends Module {
      * Next Tab press will leaf Settings Buttons
      */
     if (!this.Editor.BlockSettings.opened) {
+      /**
+       * @todo Debug the case when we set caret to some block, hovering another block
+       *       — wrong settings will be opened.
+       *       To fix it, we should refactor the Block Settings module — make it a standalone class, like the Toolbox
+       */
       this.Editor.BlockSettings.open();
     }
   }