+++ src/components/modules/toolbar/index.ts
@@ -5,8 +5,25 @@ import I18n from '../../i18n';
 import { I18nInternalNS } from '../../i18n/namespace-internal';
 import Tooltip from '../../utils/tooltip';
 import { ModuleConfig } from '../../../types-internal/module-config';
-import { EditorConfig } from '../../../../types';
-import SelectionUtils from '../../selection';
+import { BlockAPI } from '../../../../types';
+import Block from '../../block';
+import Toolbox, { ToolboxEvent } from '../../ui/toolbox';
+
+/**
+ * @todo Tab on non-empty block should open Block Settings of the hoveredBlock (not where caret is set)
+ *          - make Block Settings a standalone module
+ *
+ * @todo TESTCASE - show toggler after opening and closing the Inline Toolbar
+ * @todo TESTCASE - Click outside Editor holder should close Toolbar and Clear Focused blocks
+ * @todo TESTCASE - Click inside Editor holder should close Toolbar and Clear Focused blocks
+ * @todo TESTCASE - Click inside Redactor zone when Block Settings are opened:
+ *                  - should close Block Settings
+ *                  - should not close Toolbar
+ *                  - should move Toolbar to the clicked Block
+ * @todo TESTCASE - Toolbar should be closed on the Cross Block Selection
+ * @todo TESTCASE - Toolbar should be closed on the Rectangle Selection
+ * @todo TESTCASE - If Block Settings or Toolbox are opened, the Toolbar should not be moved by Bocks hovering
+ */
 
 /**
  * HTML Elements used for Toolbar UI
@@ -80,11 +97,22 @@ export default class Toolbar extends Module<ToolbarNodes> {
    * Tooltip utility Instance
    */
   private tooltip: Tooltip;
+
+  /**
+   * Block near which we display the Toolbox
+   */
+  private hoveredBlock: Block;
+
+  /**
+   * Toolbox class instance
+   */
+  private toolboxInstance: Toolbox;
+
   /**
    * @class
-   * @param {object} moduleConfiguration - Module Configuration
-   * @param {EditorConfig} moduleConfiguration.config - Editor's config
-   * @param {EventsDispatcher} moduleConfiguration.eventsDispatcher - Editor's event dispatcher
+   * @param moduleConfiguration - Module Configuration
+   * @param moduleConfiguration.config - Editor's config
+   * @param moduleConfiguration.eventsDispatcher - Editor's event dispatcher
    */
   constructor({ config, eventsDispatcher }: ModuleConfig) {
     super({
@@ -107,6 +135,7 @@ export default class Toolbar extends Module<ToolbarNodes> {
       actionsOpened: 'ce-toolbar__actions--opened',
 
       toolbarOpened: 'ce-toolbar--opened',
+      openedToolboxHolderModifier: 'codex-editor--toolbox-opened',
 
       // Content Zone
       plusButton: 'ce-toolbar__plus',
@@ -137,11 +166,37 @@ export default class Toolbar extends Module<ToolbarNodes> {
     return {
       hide: (): void => this.nodes.plusButton.classList.add(this.CSS.plusButtonHidden),
       show: (): void => {
-        if (this.Editor.Toolbox.isEmpty) {
+        if (this.toolboxInstance.isEmpty) {
           return;
         }
         this.nodes.plusButton.classList.remove(this.CSS.plusButtonHidden);
       },
+    };
+  }
+
+  /**
+   * Public interface for accessing the Toolbox
+   */
+  public get toolbox(): {
+    opened: boolean;
+    close: () => void;
+    open: () => void;
+    toggle: () => void;
+    flipperHasFocus: boolean;
+    } {
+    return {
+      opened: this.toolboxInstance.opened,
+      close: (): void => this.toolboxInstance.close(),
+      open: (): void => {
+        /**
+         * Set current block to cover the case when the Toolbar showed near hovered Block but caret is set to another Block.
+         */
+        this.Editor.BlockManager.currentBlock = this.hoveredBlock;
+
+        this.toolboxInstance.open();
+      },
+      toggle: (): void => this.toolboxInstance.toggle(),
+      flipperHasFocus: this.toolboxInstance.flipperHasFocus,
     };
   }
 
@@ -172,54 +227,83 @@ export default class Toolbar extends Module<ToolbarNodes> {
       this.enableModuleBindings();
     } else {
       this.destroy();
-      this.Editor.Toolbox.destroy();
       this.Editor.BlockSettings.destroy();
       this.disableModuleBindings();
     }
   }
 
   /**
-   * Move Toolbar to the Current Block
+   * Move Toolbar to the passed (or current) Block
    *
-   * @param {boolean} forceClose - force close Toolbar Settings and Toolbar
+   * @param block - block to move Toolbar near it
    */
-  public move(forceClose = true): void {
-    if (forceClose) {
-      /** Close Toolbox when we move toolbar */
-      this.Editor.Toolbox.close();
-      this.Editor.BlockSettings.close();
-    }
-
-    const currentBlock = this.Editor.BlockManager.currentBlock.holder;
+  public moveAndOpen(block: Block = this.Editor.BlockManager.currentBlock): void {
+    /**
+     * Close Toolbox when we move toolbar
+     */
+    this.toolboxInstance.close();
+    this.Editor.BlockSettings.close();
 
     /**
      * If no one Block selected as a Current
      */
-    if (!currentBlock) {
+    if (!block) {
       return;
     }
 
+    this.hoveredBlock = block;
+
+    const targetBlockHolder = block.holder;
     const { isMobile } = this.Editor.UI;
-    const blockHeight = currentBlock.offsetHeight;
-    let toolbarY = currentBlock.offsetTop;
+    const renderedContent = block.pluginsContent;
+    const renderedContentStyle = window.getComputedStyle(renderedContent);
+    const blockRenderedElementPaddingTop = parseInt(renderedContentStyle.paddingTop, 10);
+    const blockHeight = targetBlockHolder.offsetHeight;
+
+    let toolbarY;
 
     /**
-     * 1) On desktop — Toolbar at the top of Block, Plus/Toolbox moved the center of Block
-     * 2) On mobile — Toolbar at the bottom of Block
+     * On mobile — Toolbar at the bottom of Block
+     * On Desktop — Toolbar should be moved to the first line of block text
+     *              To do that, we compute the block offset and the padding-top of the plugin content
      */
-    if (!isMobile) {
-      const contentOffset = Math.floor(blockHeight / 2);
-
-      this.nodes.plusButton.style.transform = `translate3d(0, calc(${contentOffset}px - 50%), 0)`;
-      this.Editor.Toolbox.nodes.toolbox.style.transform = `translate3d(0, calc(${contentOffset}px - 50%), 0)`;
+    if (isMobile) {
+      toolbarY = targetBlockHolder.offsetTop + blockHeight;
     } else {
-      toolbarY += blockHeight;
+      toolbarY = targetBlockHolder.offsetTop + blockRenderedElementPaddingTop;
     }
 
     /**
      * Move Toolbar to the Top coordinate of Block
      */
     this.nodes.wrapper.style.transform = `translate3D(0, ${Math.floor(toolbarY)}px, 0)`;
+
+    /**
+     * Plus Button should be shown only for __empty__ __default__ block
+     */
+    if (block.tool.isDefault && block.isEmpty) {
+      this.plusButton.show();
+    } else {
+      this.plusButton.hide();
+    }
+
+    this.open();
+  }
+
+  /**
+   * Close the Toolbar
+   */
+  public close(): void {
+    if (this.Editor.ReadOnly.isEnabled) {
+      return;
+    }
+
+    this.nodes.wrapper.classList.remove(this.CSS.toolbarOpened);
+
+    /** Close components */
+    this.blockActions.hide();
+    this.toolboxInstance.close();
+    this.Editor.BlockSettings.close();
   }
 
   /**
@@ -232,9 +316,8 @@ export default class Toolbar extends Module<ToolbarNodes> {
    *                                      with closing Toolbox and Block Settings
    *                                      This flag allows to open Toolbar with Toolbox
    */
-  public open(withBlockActions = true, needToCloseToolbox = true): void {
+  private open(withBlockActions = true, needToCloseToolbox = true): void {
     _.delay(() => {
-      this.move(needToCloseToolbox);
       this.nodes.wrapper.classList.add(this.CSS.toolbarOpened);
 
       if (withBlockActions) {
@@ -243,18 +326,6 @@ export default class Toolbar extends Module<ToolbarNodes> {
         this.blockActions.hide();
       }
     }, 50)();
-  }
-
-  /**
-   * Close the Toolbar
-   */
-  public close(): void {
-    this.nodes.wrapper.classList.remove(this.CSS.toolbarOpened);
-
-    /** Close components */
-    this.blockActions.hide();
-    this.Editor.Toolbox.close();
-    this.Editor.BlockSettings.close();
   }
 
   /**
@@ -282,10 +353,11 @@ export default class Toolbar extends Module<ToolbarNodes> {
      *  - Toolbox
      */
     this.nodes.plusButton = $.make('div', this.CSS.plusButton);
-    $.append(this.nodes.plusButton, $.svg('plus', 14, 14));
-    $.append(this.nodes.content, this.nodes.plusButton);
+    $.append(this.nodes.plusButton, $.svg('plus', 16, 16));
+    $.append(this.nodes.actions, this.nodes.plusButton);
 
     this.readOnlyMutableListeners.on(this.nodes.plusButton, 'click', () => {
+      this.tooltip.hide(true);
       this.plusButtonClicked();
     }, false);
 
@@ -299,7 +371,9 @@ export default class Toolbar extends Module<ToolbarNodes> {
       textContent: '⇥ Tab',
     }));
 
-    this.tooltip.onHover(this.nodes.plusButton, tooltipContent);
+    this.tooltip.onHover(this.nodes.plusButton, tooltipContent, {
+      hidingDelay: 400,
+    });
 
     /**
      * Fill Actions Zone:
@@ -309,7 +383,7 @@ export default class Toolbar extends Module<ToolbarNodes> {
      */
     this.nodes.blockActionsButtons = $.make('div', this.CSS.blockActionsButtons);
     this.nodes.settingsToggler = $.make('span', this.CSS.settingsToggler);
-    const settingsIcon = $.svg('dots', 8, 8);
+    const settingsIcon = $.svg('dots', 16, 16);
 
     $.append(this.nodes.settingsToggler, settingsIcon);
     $.append(this.nodes.blockActionsButtons, this.nodes.settingsToggler);
@@ -319,14 +393,14 @@ export default class Toolbar extends Module<ToolbarNodes> {
       this.nodes.settingsToggler,
       I18n.ui(I18nInternalNS.ui.blockTunes.toggler, 'Click to tune'),
       {
-        placement: 'top',
+        hidingDelay: 400,
       }
     );
 
     /**
      * Appending Toolbar components to itself
      */
-    $.append(this.nodes.content, this.Editor.Toolbox.nodes.toolbox);
+    $.append(this.nodes.content, this.makeToolbox());
     $.append(this.nodes.actions, this.Editor.BlockSettings.nodes.wrapper);
 
     /**
@@ -336,10 +410,56 @@ export default class Toolbar extends Module<ToolbarNodes> {
   }
 
   /**
+   * Creates the Toolbox instance and return it's rendered element
+   */
+  private makeToolbox(): Element {
+    /**
+     * Make the Toolbox
+     */
+    this.toolboxInstance = new Toolbox({
+      api: this.Editor.API.methods,
+      tools: this.Editor.Tools.blockTools,
+    });
+
+    this.toolboxInstance.on(ToolboxEvent.Opened, () => {
+      this.Editor.UI.nodes.wrapper.classList.add(this.CSS.openedToolboxHolderModifier);
+    });
+
+    this.toolboxInstance.on(ToolboxEvent.Closed, () => {
+      this.Editor.UI.nodes.wrapper.classList.remove(this.CSS.openedToolboxHolderModifier);
+    });
+
+    this.toolboxInstance.on(ToolboxEvent.BlockAdded, ({ block }: {block: BlockAPI }) => {
+      const { BlockManager, Caret } = this.Editor;
+      const newBlock = BlockManager.getBlockById(block.id);
+
+      /**
+       * If the new block doesn't contain inputs, insert the new paragraph below
+       */
+      if (newBlock.inputs.length === 0) {
+        if (newBlock === BlockManager.lastBlock) {
+          BlockManager.insertAtEnd();
+          Caret.setToBlock(BlockManager.lastBlock);
+        } else {
+          Caret.setToBlock(BlockManager.nextBlock);
+        }
+      }
+    });
+
+    return this.toolboxInstance.make();
+  }
+
+  /**
    * Handler for Plus Button
    */
   private plusButtonClicked(): void {
-    this.Editor.Toolbox.toggle();
+    /**
+     * We need to update Current Block because user can click on the Plus Button (thanks to appearing by hover) without any clicks on editor
+     * In this case currentBlock will point last block
+     */
+    this.Editor.BlockManager.currentBlock = this.hoveredBlock;
+
+    this.toolboxInstance.toggle();
   }
 
   /**
@@ -360,7 +480,25 @@ export default class Toolbar extends Module<ToolbarNodes> {
       e.stopPropagation();
 
       this.settingsTogglerClicked();
+
+      this.toolboxInstance.close();
+
+      this.tooltip.hide(true);
     }, true);
+
+    /**
+     * Subscribe to the 'block-hovered' event
+     */
+    this.eventsDispatcher.on(this.Editor.UI.events.blockHovered, (data: {block: Block}) => {
+      /**
+       * Do not move toolbar if Block Settings or Toolbox opened
+       */
+      if (this.Editor.BlockSettings.opened || this.toolboxInstance.opened) {
+        return;
+      }
+
+      this.moveAndOpen(data.block);
+    });
   }
 
   /**
@@ -374,10 +512,16 @@ export default class Toolbar extends Module<ToolbarNodes> {
    * Clicks on the Block Settings toggler
    */
   private settingsTogglerClicked(): void {
+    /**
+     * We need to update Current Block because user can click on toggler (thanks to appearing by hover) without any clicks on editor
+     * In this case currentBlock will point last block
+     */
+    this.Editor.BlockManager.currentBlock = this.hoveredBlock;
+
     if (this.Editor.BlockSettings.opened) {
       this.Editor.BlockSettings.close();
     } else {
-      this.Editor.BlockSettings.open();
+      this.Editor.BlockSettings.open(this.hoveredBlock);
     }
   }
 
@@ -385,7 +529,7 @@ export default class Toolbar extends Module<ToolbarNodes> {
    * Draws Toolbar UI
    *
    * Toolbar contains BlockSettings and Toolbox.
-   * Thats why at first we draw its components and then Toolbar itself
+   * That's why at first we draw its components and then Toolbar itself
    *
    * Steps:
    *  - Make Toolbar dependent components like BlockSettings, Toolbox and so on
@@ -399,11 +543,6 @@ export default class Toolbar extends Module<ToolbarNodes> {
     this.Editor.BlockSettings.make();
 
     /**
-     * Make Toolbox
-     */
-    this.Editor.Toolbox.make();
-
-    /**
      * Make Toolbar
      */
     this.make();
@@ -415,6 +554,9 @@ export default class Toolbar extends Module<ToolbarNodes> {
    */
   private destroy(): void {
     this.removeAllNodes();
+    if (this.toolboxInstance) {
+      this.toolboxInstance.destroy();
+    }
     this.tooltip.destroy();
   }
 }